Refactor main_gui
Simplify, cleanup, document.
